moved find_key_in_dictionary to utils.helpers
